splitting out headder file
